Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,9 +3,9 @@
 
 
 int main() {
-    int x,y,z;
+    float x,y,z;
     scanf("%d%d%d",&x,&y,&z);
-    int average=(x+y+z/3);
+    float average=((x+y+z)/3);
     printf("Average: %d",average);
  
     return 0;
